Improve mouse turning and sensitivity values
This commit outlines a change to syngen that would require a new release. Essentially the call to setTimeout() within syngen.input.mouse.update() will sometimes occur before input is processed by the app. Instead we might provide a method to reset this after it's been read, or modify the module to expose a new copy of the data each frame.